Refactor WorkflowEditor and useIopeer
