+++ front/src/components/features/index.js
@@ -1,2 +1,3 @@
 export { default as Agents } from './Agents/Agents';
 export { default as Marketplace } from './Marketplace/Marketplace';
+export { WorkflowEditor } from './WorkflowEditor';
