+++ front/src/hooks/iopeer/useUIState.js
@@ -1,0 +1,6 @@
+export const deriveUIState = ({ connectionStatus, agents }) => ({
+  isConnected: connectionStatus === 'connected',
+  isConnecting: connectionStatus === 'connecting',
+  isFailed: connectionStatus === 'failed',
+  hasAgents: agents.length > 0,
+});
